Added RJ45 LED-Supprt
Removed old and unused "green_led" code.
(Which was used during testing,
later on the green LED was disconnected.)

Replaced the old "green_led" with the RJ45 green LED.
Can be used for demonstrations and BUS tests.

Added yellow RJ45 LED to activate while processing data.
This can be used to show a message was received successfully.